tests/utils: Make "wrap" easier to call by hand
When doing some tests by hand, some times it can be convenient to call
"wrap" from the command line.

Currently this is not as easy, because there are some assumptions about
paths that get in the way.

This patch removes that annoyance by making paths relative to the
utility itself.

As a side-effect, update travis-ci's configuration to be based on
xenial, as this change needs "realpath" which has been a part of
coreutils for a few years.